# services/api-gateway/internal/docs/openapi/openapi-organization.yaml
openapi: 3.0.3
info:
  title: CivicOS Organization Service
  version: 1.0.0
  description: |
    Organizations (registry + membership), announcements, projects, issue
    assignments ("receive reports"), and progress updates ("respond publicly",
    "update progress").

    All routes are proxied by the API gateway at `/api/v1/*`. The service
    itself mounts them under `/v1/*`.

    ## Authorization model

    Roles allowed to **create** a top-level org from scratch: `GOVERNMENT_ADMIN`,
    `PLATFORM_ADMIN`, `NGO`. Once created, the org's internal role
    (`OWNER` / `ADMIN` / `STAFF`) governs edits — not the platform JWT role.

    **Content-authorship writes are strictly org-only.** A `PLATFORM_ADMIN`
    who is not a member of the org **cannot** create, edit, publish, or
    delete announcements / projects / assignments / consultations on the
    org's behalf. Attribution belongs to the organization itself.

    Platform admins retain two narrow powers:

    - **Emergency close/archive** — `PLATFORM_ADMIN` can close a published
      consultation or archive an announcement without joining the org first.
      Written to the audit log with the actor's identity.
    - **Oversight reads** — `PLATFORM_ADMIN` can view responses, analytics,
      and org-only content (draft announcements, assignment inboxes) for
      any org.

    Anything beyond these two levers requires the platform admin to be
    added as a member of the target org first.
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrgKind:
      type: string
      enum: [GOVERNMENT, AGENCY, NGO, UTILITY, OTHER]
    OrgJurisdiction:
      type: string
      enum: [NATIONAL, STATE, LGA, COMMUNITY]
    OrgMemberRole:
      type: string
      enum: [OWNER, ADMIN, STAFF]
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      security: []
      summary: List organizations
      parameters:
        - in: query
          name: kind
          schema: { $ref: '#/components/schemas/OrgKind' }
        - in: query
          name: jurisdiction
          schema: { $ref: '#/components/schemas/OrgJurisdiction' }
        - in: query
          name: state
          schema: { type: string }
        - in: query
          name: lga
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Organizations matching the filter.
          content:
            application/json:
              examples:
                default:
                  value:
                    success: true
                    data:
                      organizations:
                        - id: 9a8b7c6d-...
                          name: Enugu Water Corporation
                          slug: enugu-water-corp
                          kind: UTILITY
                          jurisdiction: STATE
                          state: Enugu
                          verified: true
                          memberCount: 12
                          projectCount: 4
                          assignmentCount: 27
                          createdById: 3f1a...
                          createdAt: 2026-06-01T09:00:00Z
                          updatedAt: 2026-06-01T09:00:00Z
    post:
      tags: [Organizations]
      summary: Create a new organization (GOVERNMENT_ADMIN, PLATFORM_ADMIN, NGO)
      description: The creator is added as `OWNER` automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, kind, jurisdiction]
              properties:
                name: { type: string, minLength: 2 }
                slug: { type: string, description: lowercase-hyphenated }
                kind: { $ref: '#/components/schemas/OrgKind' }
                jurisdiction: { $ref: '#/components/schemas/OrgJurisdiction' }
                state: { type: string, nullable: true }
                lga: { type: string, nullable: true }
                description: { type: string, nullable: true }
                logoUrl: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
                phone: { type: string, nullable: true }
                website: { type: string, format: uri, nullable: true }
            examples:
              default:
                value:
                  {
                    name: Enugu Water Corporation,
                    slug: enugu-water-corp,
                    kind: UTILITY,
                    jurisdiction: STATE,
                    state: Enugu,
                  }
      responses:
        '201':
          description: Organization created; audit-logged.
        '400':
          description: Invalid kind, jurisdiction, or slug format.
        '409':
          description: Slug already taken.
          content:
            application/json:
              examples:
                default:
                  {
                    value:
                      { success: false, code: SLUG_TAKEN, message: That slug is already in use },
                  }
  /organizations/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Organizations]
      security: []
      summary: Get an organization
      responses:
        '200': { description: Organization detail. }
        '404':
          description: Not found.
          content:
            application/json:
              examples:
                default:
                  {
                    value: { success: false, code: ORG_NOT_FOUND, message: Organization not found },
                  }
    patch:
      tags: [Organizations]
      summary: Update an organization (org admin or PLATFORM_ADMIN)
      description: |
        All fields are optional. Toggling `verified` writes a distinct
        `org.verified` / `org.unverified` audit entry (separate from the
        plain `org.updated` action) because the verified badge is a
        citizen-facing trust signal.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                kind: { $ref: '#/components/schemas/OrgKind' }
                jurisdiction: { $ref: '#/components/schemas/OrgJurisdiction' }
                state: { type: string, nullable: true }
                lga: { type: string, nullable: true }
                description: { type: string, nullable: true }
                logoUrl: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
                phone: { type: string, nullable: true }
                website: { type: string, format: uri, nullable: true }
                verified:
                  type: boolean
                  description: Platform admins only.
      responses:
        '200': { description: Updated organization. }

  # ────────────────────────── MEMBERS ──────────────────────────
  /organizations/{id}/members:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Members]
      security: []
      summary: List organization members
      responses:
        '200':
          description: Members.
          content:
            application/json:
              examples:
                default:
                  value:
                    success: true
                    data:
                      members:
                        - id: m1a2b3c4-...
                          organizationId: 9a8b7c6d-...
                          userId: 3f1a...
                          userName: Ada Nnamdi
                          userRole: NGO
                          role: OWNER
                          joinedAt: 2026-06-01T09:00:00Z
    post:
      tags: [Members]
      summary: Add a member (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, userName, userRole, role]
              properties:
                userId: { type: string, format: uuid }
                userName: { type: string }
                userRole: { type: string, description: Platform role at time of add. }
                role: { $ref: '#/components/schemas/OrgMemberRole' }
            examples:
              default:
                { value: { userId: 3f1a..., userName: Ada Nnamdi, userRole: NGO, role: ADMIN } }
      responses:
        '201': { description: Member added; audit-logged. }
  /organizations/{id}/members/{userId}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
      - in: path
        name: userId
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Members]
      summary: Change a member's role (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/OrgMemberRole' }
      responses:
        '200': { description: Role updated; audit-logged. }
    delete:
      tags: [Members]
      summary: Remove a member (org admin)
      responses:
        '200': { description: Member removed; audit-logged. }

  # ─────────────────────── ANNOUNCEMENTS ───────────────────────
...